chore(user): improve UserService logic, document user routes, and fix 'name' field in clients table

+++ app/services/user_service.py
@@ -2,11 +2,21 @@ from sqlalchemy.orm import Session
 from fastapi import HTTPException, status
 from typing import Optional
 from passlib.context import CryptContext
+from app.enums.role_enum import RoleEnum
 from app.models.client_model import ClientModel
 from app.schemas.user_schema import UserBase, UserCreate, UserUpdate
 from app.utils.db_exceptions import handle_db_exceptions
 
 class UserService:
+    ACCESS_DENIED = "Access denied."
+    USER_NOT_FOUND = "User not found."
+    CANNOT_DELETE_SELF = "You cannot delete your own user."
+    ACCESS_DENIED_UPDATE = "Access denied. You can only update your own data unless you are an administrator."
+    ADMIN_ROLE_REQUIRED = "Access denied. Only administrators can update user roles."
+    EMAIL_ALREADY_REGISTERED = "Email already registered."
+    CPF_ALREADY_REGISTERED = "CPF already registered."
+    CANNOT_DELETE_OWN_USER = "You cannot delete your own user."
+
     def __init__(self, client_model: ClientModel):
         self.client_model = client_model
         self.pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
@@ -30,39 +40,41 @@ class UserService:
         return query.offset(skip).limit(limit).all()
 
     @handle_db_exceptions
-    def get_user_by_id(self, db: Session, user_id: int) -> ClientModel:
-        user = db.query(self.client_model).filter(self.client_model.id == user_id).first()
+    def get_user_by_id(
+        self,
+        db: Session,
+        user_id: int,
+        current_user: ClientModel
+    ) -> ClientModel:
+        if current_user.role != 'admin' and current_user.id != user_id:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail=self.ACESS_DENIED
+            )
+        
+        user = db.query(self.client_model)\
+                .filter(self.client_model.id == user_id)\
+                .first()
+        
         if not user:
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found."
+                detail=self.USER_NOT_FOUND
             )
+        
         return user
     
     @handle_db_exceptions
     def get_user_by_email(self, db: Session, email: str) -> Optional[ClientModel]:
-        return db.query(self.client_model).filter(self.client_model.email == email).first()
-
+        return db.query(self.client_model)\
+                 .filter(self.client_model.email == email)\
+                 .first()
+    
     @handle_db_exceptions
     def get_user_by_cpf(self, db: Session, cpf: str) -> Optional[ClientModel]:
-        return db.query(self.client_model).filter(self.client_model.cpf == cpf).first()
-
-    @handle_db_exceptions
-    def check_unique_email(self, db: Session, email: str, exclude_user_id: Optional[int] = None):
-        user = self.get_user_by_email(db, email)
-        if user and (exclude_user_id is None or user.id != exclude_user_id):
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail="Email already registered."
-            )
-    @handle_db_exceptions
-    def check_unique_cpf(self, db: Session, cpf: str, exclude_user_id: Optional[int] = None):
-        user = self.get_user_by_cpf(db, cpf)
-        if user and (exclude_user_id is None or user.id != exclude_user_id):
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail="CPF already registered."
-            )
+        return db.query(self.client_model)\
+                 .filter(self.client_model.cpf == cpf)\
+                 .first()
         
     @handle_db_exceptions
     def create_user(self, db: Session, user_data: UserCreate) -> ClientModel:
@@ -92,14 +104,20 @@ class UserService:
         user_data: UserUpdate,
         current_user: ClientModel
     ) -> ClientModel:
-        if current_user.id != user_id and current_user.role != "ADMIN":
+        if current_user.id != user_id and current_user.role != 'admin':
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
-                detail="Access denied. You can only update your own data unless you are an administrator."
+                detail=self.ACCESS_DENIED_UPDATE
             )
 
-        user = self.get_user_by_id(db, user_id)
+        user = self.get_user_by_id(db, user_id, current_user)
         update_fields = user_data.model_dump(exclude_unset=True)
+
+        if "role" in update_fields and current_user.role != 'admin':
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail=self.ADMIN_ROLE_REQUIRED
+            )
 
         if "email" in update_fields:
             self.check_unique_email(db, update_fields["email"], exclude_user_id=user_id)
@@ -117,13 +135,56 @@ class UserService:
         return user
 
     @handle_db_exceptions
-    def delete_user(self, db: Session, user_id: int) -> None:
-        user = self.get_user_by_id(db, user_id)
+    def delete_user(self, db: Session, user_id: int, current_user) -> None:
+        if user_id == current_user.id:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail=self.CANNOT_DELETE_OWN_USER
+            )
+        
+        user = self.get_user_by_id(db, user_id, current_user)
+        
         db.delete(user)
         db.commit()
+
+
+    @handle_db_exceptions
+    def check_unique_email(
+        self,
+        db: Session,
+        email: str,
+        exclude_user_id: Optional[int] = None
+    ):
+        user = self.get_user_by_email(db, email)
+        if user and (exclude_user_id is None or user.id != exclude_user_id):
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail=self.EMAIL_ALREADY_REGISTERED
+            )
+        
+    @handle_db_exceptions
+    def check_unique_cpf(
+        self,
+        db: Session,
+        cpf: str,
+        exclude_user_id: Optional[int] = None
+    ):
+        user = self.get_user_by_cpf(db, cpf)
+        if user and (exclude_user_id is None or user.id != exclude_user_id):
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail=self.CPF_ALREADY_REGISTERED
+            )
 
     def hash_password(self, password: str) -> str:
         return self.pwd_context.hash(password)
 
-    def verify_password(self, plain_password: str, hashed_password: str) -> bool:
-        return self.pwd_context.verify(plain_password, hashed_password)
+    def verify_password(
+        self,
+        plain_password: str,
+        hashed_password: str
+    ) -> bool:
+        return self.pwd_context.verify(
+            plain_password, hashed_password
+        )
+
